requirements.txt
flask
python-dotenv
psycopg
psycopg-binary
bootstrap-flask
WTForms Flask-WTF
flask-login